refactor(client): ViewContentDetail에서 View와 Controller를 구분했습니다. 추가적으로 api의 명칭을 통일했씁니다. 행동_타켓_Redux (ex : ChangeDataRedux)
