+++ client/src/api/InitiateDataRedux.js
@@ -1,10 +1,8 @@
-import { useDispatch } from "react-redux";
 import FetchContent from "./FetchContent";
 import { InitContent } from "../redux/BlogContent";
 
-const initiateData = async () => {
-  const dispatch = useDispatch();
+const initiateContentData = async (dispatch) => {
   const newArr = await FetchContent();
   dispatch(InitContent(newArr));
 };
-export default initiateData;
+export default initiateContentData;
